A4A > Reports: refactor the build report flow
