+++ client/a8c-for-agencies/sections/reports/primary/build-report/steps/types.ts
@@ -1,0 +1,11 @@
+import type {
+	BuildReportFormData,
+	BuildReportState,
+	UseDuplicateReportFormDataHandlers,
+} from '../../../types';
+
+export interface StepProps {
+	formData: BuildReportFormData;
+	state: BuildReportState;
+	handlers: UseDuplicateReportFormDataHandlers;
+}
